@@ -1,20 +1,16 @@
 <Project Sdk="Microsoft.NET.Sdk.Worker">
-	<PropertyGroup>
-		<TargetFramework>netcoreapp3.1</TargetFramework>
-	</PropertyGroup>
 
-	<ItemGroup>
-		<PackageReference Include="Hub.Shared.HostedServices.ServiceBusQueue" Version="3.3" />
-		<PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="8.1.0" />
-		<PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="3.1.8" />
-	</ItemGroup>
+    <PropertyGroup>
+        <TargetFramework>net6.0</TargetFramework>
+    </PropertyGroup>
 
-	<ItemGroup>
-		<ProjectReference Include="..\Spreadsheet.Shared\Spreadsheet.Shared.csproj" />
-		<ProjectReference Include="..\Spreadsheet.Data\Spreadsheet.Data.csproj" />
-		<ProjectReference Include="..\Spreadsheet.Integration\Spreadsheet.Integration.csproj" />
-		<ProjectReference Include="..\Spreadsheet.Providers\Spreadsheet.Providers.csproj" />
-		<ProjectReference Include="..\Spreadsheet.Services\Spreadsheet.Services.csproj" />
-	</ItemGroup>
+    <ItemGroup>
+      <ProjectReference Include="..\..\hub-shared\Hub.Shared.HostedServices\Hub.Shared.HostedServices.ServiceBusQueue\Hub.Shared.HostedServices.ServiceBusQueue.csproj" />
+      <ProjectReference Include="..\Spreadsheet.Shared\Spreadsheet.Shared.csproj" />
+      <ProjectReference Include="..\Spreadsheet.Data\Spreadsheet.Data.csproj" />
+      <ProjectReference Include="..\Spreadsheet.Integration\Spreadsheet.Integration.csproj" />
+      <ProjectReference Include="..\Spreadsheet.Providers\Spreadsheet.Providers.csproj" />
+      <ProjectReference Include="..\Spreadsheet.Services\Spreadsheet.Services.csproj" />
+    </ItemGroup>
 
 </Project>
