@@ -1,13 +1,10 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 	<PropertyGroup>
-		<TargetFramework>netcoreapp3.1</TargetFramework>
+		<TargetFramework>net6.0</TargetFramework>
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Hub.Shared.Web.Api" Version="3.3" />
-	</ItemGroup>
-
-	<ItemGroup>
+		<ProjectReference Include="..\..\hub-shared\Hub.Shared.Web\Hub.Shared.Web.Api\Hub.Shared.Web.Api.csproj" />
 		<ProjectReference Include="..\Spreadsheet.Data\Spreadsheet.Data.csproj" />
 	</ItemGroup>
 
