@@ -1,17 +1,16 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 	<PropertyGroup>
-		<TargetFramework>net5.0</TargetFramework>
+		<TargetFramework>net6.0</TargetFramework>
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Blazored.FluentValidation" Version="2.0.1" />
-		<PackageReference Include="FluentValidation.AspNetCore" Version="10.2.3" />
-		<PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.17.0" />
-		<PackageReference Include="Hub.Shared.Logging" Version="3.3" />
-		<PackageReference Include="Hub.Shared.Web.BlazorServer" Version="3.3" />
+		<PackageReference Include="Blazored.FluentValidation" Version="2.0.3" />
+		<PackageReference Include="FluentValidation.AspNetCore" Version="10.3.5" />
 	</ItemGroup>
 
 	<ItemGroup>
+		<ProjectReference Include="..\..\hub-shared\Hub.Shared.Logging\Hub.Shared.Logging.csproj" />
+		<ProjectReference Include="..\..\hub-shared\Hub.Shared.Web\Hub.Shared.Web.BlazorServer\Hub.Shared.Web.BlazorServer.csproj" />
 		<ProjectReference Include="..\Spreadsheet.Data\Spreadsheet.Data.csproj" />
 		<ProjectReference Include="..\Spreadsheet.Providers\Spreadsheet.Providers.csproj" />
 		<ProjectReference Include="..\Spreadsheet.Services\Spreadsheet.Services.csproj" />
