@@ -2,14 +2,13 @@ using AutoMapper;
 using Spreadsheet.Data.Documents;
 using Spreadsheet.Data.Dto;
 
-namespace Spreadsheet.Data.AutoMapper
+namespace Spreadsheet.Data.AutoMapper;
+
+public class SpreadsheetMetadataMapperProfile : Profile
 {
-    public class SpreadsheetMetadataMapperProfile : Profile
+    public SpreadsheetMetadataMapperProfile()
     {
-        public SpreadsheetMetadataMapperProfile()
-        {
-            CreateMap<SpreadsheetMetadata, SpreadsheetMetadataDto>()
-                .ReverseMap();
-        }
+        CreateMap<SpreadsheetMetadata, SpreadsheetMetadataDto>()
+            .ReverseMap();
     }
 }
