@@ -2,18 +2,17 @@
 using Microsoft.EntityFrameworkCore;
 using Spreadsheet.Data.Entities;
 
-namespace Spreadsheet.Data
+namespace Spreadsheet.Data;
+
+public class SpreadsheetDbContext : HubDbContext
 {
-    public class SpreadsheetDbContext : HubDbContext
-    {
-        public SpreadsheetDbContext(DbContextOptions<SpreadsheetDbContext> options) : base(options) { }
+    public SpreadsheetDbContext(DbContextOptions<SpreadsheetDbContext> options) : base(options) { }
         
-        protected override void OnModelCreating(ModelBuilder builder)
-        {
-            base.OnModelCreating(builder);
+    protected override void OnModelCreating(ModelBuilder builder)
+    {
+        base.OnModelCreating(builder);
             
-            builder.Entity<BillingAccountTransaction>()
-                .ToTable(schema: "dbo", name: "BillingAccountPayment");
-        }
+        builder.Entity<BillingAccountTransaction>()
+            .ToTable(schema: "dbo", name: "BillingAccountPayment");
     }
 }
