@@ -3,15 +3,14 @@ using Spreadsheet.Integration.Dto.Spreadsheet.Budget.Tabs;
 using Spreadsheet.Providers;
 using Spreadsheet.Services;
 
-namespace Spreadsheet.HostedServices.ServiceBusQueueHost.Commands
+namespace Spreadsheet.HostedServices.ServiceBusQueueHost.Commands;
+
+public class UpdateBillingAccountTransactionsCommand : UpdateTabCommandBase<BillingAccountTab>
 {
-    public class UpdateBillingAccountTransactionsCommand : UpdateTabCommandBase<BillingAccountTab>
+    public UpdateBillingAccountTransactionsCommand(ITabWriterService<BillingAccountTab> tabWriterService,
+        ITabDataProvider<BillingAccountTab> billingAccountPaymentsDataProvider) : base(billingAccountPaymentsDataProvider, tabWriterService)
     {
-        public UpdateBillingAccountTransactionsCommand(ITabWriterService<BillingAccountTab> tabWriterService,
-            ITabDataProvider<BillingAccountTab> billingAccountPaymentsDataProvider) : base(billingAccountPaymentsDataProvider, tabWriterService)
-        {
-        }
-        
-        public override string Trigger => QueueNames.SbankenTransactionsUpdated;
     }
+        
+    public override string Trigger => QueueNames.SbankenTransactionsUpdated;
 }
