@@ -3,15 +3,14 @@ using Spreadsheet.Integration.Dto.Spreadsheet.Budget.Tabs;
 using Spreadsheet.Providers;
 using Spreadsheet.Services;
 
-namespace Spreadsheet.HostedServices.ServiceBusQueueHost.Commands
-{
-    public class UpdateCoinbaseProAccountsCommand : UpdateTabCommandBase<CoinbaseProAccountsTab>
-    {
-        public UpdateCoinbaseProAccountsCommand(ITabWriterService<CoinbaseProAccountsTab> tabWriterService,
-            ITabDataProvider<CoinbaseProAccountsTab> tabDataProvider) : base(tabDataProvider, tabWriterService)
-        {
-        }
+namespace Spreadsheet.HostedServices.ServiceBusQueueHost.Commands;
 
-        public override string Trigger => QueueNames.CoinbaseProAccountsUpdated;
+public class UpdateCoinbaseProAccountsCommand : UpdateTabCommandBase<CoinbaseProAccountsTab>
+{
+    public UpdateCoinbaseProAccountsCommand(ITabWriterService<CoinbaseProAccountsTab> tabWriterService,
+        ITabDataProvider<CoinbaseProAccountsTab> tabDataProvider) : base(tabDataProvider, tabWriterService)
+    {
     }
+
+    public override string Trigger => QueueNames.CoinbaseProAccountsUpdated;
 }
