@@ -4,25 +4,24 @@ using System.Threading.Tasks;
 using JetBrains.Annotations;
 using Spreadsheet.Integration.Dto;
 
-namespace Spreadsheet.Integration
+namespace Spreadsheet.Integration;
+
+public interface ICoinbaseApiConnector : IBankApiConnector
 {
-    public interface ICoinbaseApiConnector : IBankApiConnector
+    Task<IList<ExchangeRateDto>> GetExchangeRates();
+}
+
+[UsedImplicitly]
+public class CoinbaseApiConnector : BankApiConnector, ICoinbaseApiConnector
+{
+    private const string ExchangeRatesPath = "/api/exchangerates";
+
+    public CoinbaseApiConnector(HttpClient httpClient) : base(httpClient, "CoinbaseApi")
     {
-        Task<IList<ExchangeRateDto>> GetExchangeRates();
     }
 
-    [UsedImplicitly]
-    public class CoinbaseApiConnector : BankApiConnector, ICoinbaseApiConnector
+    public async Task<IList<ExchangeRateDto>> GetExchangeRates()
     {
-        private const string ExchangeRatesPath = "/api/exchangerates";
-
-        public CoinbaseApiConnector(HttpClient httpClient) : base(httpClient, "CoinbaseApi")
-        {
-        }
-
-        public async Task<IList<ExchangeRateDto>> GetExchangeRates()
-        {
-            return await Get<IList<ExchangeRateDto>>(ExchangeRatesPath);
-        }
+        return await Get<IList<ExchangeRateDto>>(ExchangeRatesPath);
     }
 }
