@@ -1,17 +1,16 @@
 using System.Net.Http;
 using JetBrains.Annotations;
 
-namespace Spreadsheet.Integration
-{
-    public interface ICoinbaseProApiConnector : IBankApiConnector
-    {
-    }
+namespace Spreadsheet.Integration;
 
-    [UsedImplicitly]
-    public class CoinbaseProApiConnector : BankApiConnector, ICoinbaseProApiConnector
+public interface ICoinbaseProApiConnector : IBankApiConnector
+{
+}
+
+[UsedImplicitly]
+public class CoinbaseProApiConnector : BankApiConnector, ICoinbaseProApiConnector
+{
+    public CoinbaseProApiConnector(HttpClient httpClient) : base(httpClient, "CoinbaseProApi")
     {
-        public CoinbaseProApiConnector(HttpClient httpClient) : base(httpClient, "CoinbaseProApi")
-        {
-        }
     }
 }
