@@ -2,12 +2,11 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Spreadsheet.Integration.Dto;
 
-namespace Spreadsheet.Integration
-{
-    public interface IBankApiConnector
-    {
-        string FriendlyApiName { get; }
+namespace Spreadsheet.Integration;
 
-        Task<IList<AccountDto>> GetAccounts();
-    }
+public interface IBankApiConnector
+{
+    string FriendlyApiName { get; }
+
+    Task<IList<AccountDto>> GetAccounts();
 }
