@@ -4,26 +4,25 @@ using System.Threading.Tasks;
 using JetBrains.Annotations;
 using Spreadsheet.Integration.Dto;
 
-namespace Spreadsheet.Integration
+namespace Spreadsheet.Integration;
+
+public interface ISbankenApiConnector : IBankApiConnector
 {
-    public interface ISbankenApiConnector : IBankApiConnector
+    Task<IList<TransactionDto>> GetBillingAccountTransactions(int ageInDays);
+}
+
+[UsedImplicitly]
+public class SbankenApiConnector : BankApiConnector, ISbankenApiConnector
+{
+    private const string TransactionsPath = "/api/transactions";
+
+    public SbankenApiConnector(HttpClient httpClient) : base(httpClient, "SbankenApi")
     {
-        Task<IList<TransactionDto>> GetBillingAccountTransactions(int ageInDays);
     }
 
-    [UsedImplicitly]
-    public class SbankenApiConnector : BankApiConnector, ISbankenApiConnector
+    public async Task<IList<TransactionDto>> GetBillingAccountTransactions(int ageInDays)
     {
-        private const string TransactionsPath = "/api/transactions";
-
-        public SbankenApiConnector(HttpClient httpClient) : base(httpClient, "SbankenApi")
-        {
-        }
-
-        public async Task<IList<TransactionDto>> GetBillingAccountTransactions(int ageInDays)
-        {
-            return await Get<IList<TransactionDto>>(TransactionsPath,
-                $"ageInDays={ageInDays}&accountName=Regningsbetaling");
-        }
+        return await Get<IList<TransactionDto>>(TransactionsPath,
+            $"ageInDays={ageInDays}&accountName=Regningsbetaling");
     }
 }
