@@ -1,10 +1,9 @@
-namespace Spreadsheet.Shared.Constants
+namespace Spreadsheet.Shared.Constants;
+
+public static class QueueNames
 {
-    public static class QueueNames
-    {
-        public static readonly string SbankenTransactionsUpdated = "sbankentransactionsupdated";
-        public static readonly string CoinbaseProAccountsUpdated = "coinbaseproaccountsupdated";
-        public static readonly string ExchangeRatesUpdated = "coinbaseexchangeratesupdated";
-        public static readonly string SbankenAccountsUpdated = "sbankenaccountsupdated";
-    }
+    public static readonly string SbankenTransactionsUpdated = "sbankentransactionsupdated";
+    public static readonly string CoinbaseProAccountsUpdated = "coinbaseproaccountsupdated";
+    public static readonly string ExchangeRatesUpdated = "coinbaseexchangeratesupdated";
+    public static readonly string SbankenAccountsUpdated = "sbankenaccountsupdated";
 }
