@@ -1,8 +1,7 @@
-namespace Spreadsheet.Shared.Constants
+namespace Spreadsheet.Shared.Constants;
+
+public static class SettingConstants
 {
-    public static class SettingConstants
-    {
-        public static readonly string GoogleCertificate = "GoogleCertificate";
-        public static readonly string GoogleServiceAccountEmail = "GoogleServiceAccountEmail";
-    }
+    public static readonly string GoogleCertificate = "GoogleCertificate";
+    public static readonly string GoogleServiceAccountEmail = "GoogleServiceAccountEmail";
 }
