@@ -1,11 +1,10 @@
-namespace Spreadsheet.Shared.Constants
+namespace Spreadsheet.Shared.Constants;
+
+public static class SpreadsheetTabMetadataConstants
 {
-    public static class SpreadsheetTabMetadataConstants
-    {
-        public static readonly string SbankenAccountsTabName = "SBANKEN_ACCOUNTS";
-        public static readonly string CoinbaseProAccountsTabName = "COINBASE_PRO_ACCOUNTS";
-        public static readonly string ResultAndSavingTabName = "Resultat og sparing";
-        public static readonly string BillingAccountTabName = "API_REGNINGSKONTO";
-        public static readonly string ExchangeRatesTabName = "EXCHANGE_RATES";
-    }
+    public static readonly string SbankenAccountsTabName = "SBANKEN_ACCOUNTS";
+    public static readonly string CoinbaseProAccountsTabName = "COINBASE_PRO_ACCOUNTS";
+    public static readonly string ResultAndSavingTabName = "Resultat og sparing";
+    public static readonly string BillingAccountTabName = "API_REGNINGSKONTO";
+    public static readonly string ExchangeRatesTabName = "EXCHANGE_RATES";
 }
