@@ -1,12 +1,11 @@
 using System.Globalization;
 
-namespace Spreadsheet.Shared.Extensions
+namespace Spreadsheet.Shared.Extensions;
+
+public static class DecimalExtensions
 {
-    public static class DecimalExtensions
+    public static string ToComma(this decimal number) 
     {
-        public static string ToComma(this decimal number) 
-        {
-            return number.ToString(CultureInfo.CurrentCulture).Replace(".",",");
-        }
+        return number.ToString(CultureInfo.CurrentCulture).Replace(".",",");
     }
 }
