@@ -3,16 +3,15 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Spreadsheet.Data;
 
-namespace Spreadsheet.Web.Api
-{
-    public class DependencyRegistrationFactory : DependencyRegistrationFactory<SpreadsheetDbContext>
-    {
-        public DependencyRegistrationFactory() : base("SQL_DB_SPREADSHEET", "Spreadsheet.Data")
-        {
-        }
+namespace Spreadsheet.Web.Api;
 
-        protected override void AddDomainDependencies(IServiceCollection serviceCollection, IConfiguration configuration)
-        {
-        }
+public class DependencyRegistrationFactory : DependencyRegistrationFactory<SpreadsheetDbContext>
+{
+    public DependencyRegistrationFactory() : base("SQL_DB_SPREADSHEET", "Spreadsheet.Data")
+    {
+    }
+
+    protected override void AddDomainDependencies(IServiceCollection serviceCollection, IConfiguration configuration)
+    {
     }
 }
