@@ -1,4 +1,3 @@
-using AutoMapper;
 using FluentValidation;
 using FluentValidation.AspNetCore;
 using Microsoft.Extensions.Configuration;
@@ -10,29 +9,28 @@ using Spreadsheet.Providers;
 using Spreadsheet.Services;
 using Spreadsheet.Web.WebApp.Validation;
 
-namespace Spreadsheet.Web.WebApp
+namespace Spreadsheet.Web.WebApp;
+
+public class DependencyRegistrationFactory : Hub.Shared.Web.BlazorServer.DependencyRegistrationFactory
 {
-    public class DependencyRegistrationFactory : Hub.Shared.Web.BlazorServer.DependencyRegistrationFactory
+    protected override void AddBlazorExtras(IServiceCollection serviceCollection, IConfiguration configuration)
     {
-        protected override void AddBlazorExtras(IServiceCollection serviceCollection, IConfiguration configuration)
-        {
-        }
+    }
 
-        protected override void AddHttpClients(IServiceCollection serviceCollection, IConfiguration configuration)
-        {
-        }
+    protected override void AddHttpClients(IServiceCollection serviceCollection, IConfiguration configuration)
+    {
+    }
 
-        protected override void AddDomainDependencies(IServiceCollection serviceCollection, IConfiguration configuration)
-        {
-            serviceCollection.AddTransient<ISpreadsheetCosmosDb, SpreadsheetCosmosDb>();
-            serviceCollection.AddTransient<ISpreadsheetMetadataProvider, SpreadsheetMetadataProvider>();
-            serviceCollection.AddTransient<ISpreadsheetMetadataService, SpreadsheetMetadataService>();
-            serviceCollection.AddSingleton<State>();
+    protected override void AddDomainDependencies(IServiceCollection serviceCollection, IConfiguration configuration)
+    {
+        serviceCollection.AddTransient<ISpreadsheetCosmosDb, SpreadsheetCosmosDb>();
+        serviceCollection.AddTransient<ISpreadsheetMetadataProvider, SpreadsheetMetadataProvider>();
+        serviceCollection.AddTransient<ISpreadsheetMetadataService, SpreadsheetMetadataService>();
+        serviceCollection.AddSingleton<State>();
 
-            serviceCollection.AddAutoMapper(c => { c.AddSpreadsheetProfiles(); });
+        serviceCollection.AddAutoMapper(c => { c.AddSpreadsheetProfiles(); });
 
-            serviceCollection.AddFluentValidation();
-            serviceCollection.AddTransient<IValidator<SpreadsheetMetadataDto>, SpreadsheetMetadataValidator>();
-        }
+        serviceCollection.AddFluentValidation();
+        serviceCollection.AddTransient<IValidator<SpreadsheetMetadataDto>, SpreadsheetMetadataValidator>();
     }
 }
