@@ -1,26 +1,30 @@
 using System;
+using System.Globalization;
 using Microsoft.AspNetCore.Components;
 
-namespace Spreadsheet.Web.WebApp.Helpers
+namespace Spreadsheet.Web.WebApp.Helpers;
+
+public static class InputFormatter
 {
-    public static class InputFormatter
+    public static string ToUpperCase(ChangeEventArgs changeEventArgs)
     {
-        public static string ToUpperCase(ChangeEventArgs changeEventArgs)
+        if (changeEventArgs.Value is not string text)
         {
-            if (changeEventArgs.Value is not string text)
-                throw new ArgumentException("Input type was not string", nameof(changeEventArgs));
-
-            changeEventArgs.Value = text.ToUpper();
-
-            return Trim(changeEventArgs);
+            throw new ArgumentException("Input type was not string", nameof(changeEventArgs));
         }
 
-        public static string Trim(ChangeEventArgs changeEventArgs)
-        {
-            if (changeEventArgs.Value is not string text)
-                throw new ArgumentException("Input type was not string", nameof(changeEventArgs));
+        changeEventArgs.Value = text.ToUpper(CultureInfo.InvariantCulture);
 
-            return text.Replace(" ", "");
+        return Trim(changeEventArgs);
+    }
+
+    public static string Trim(ChangeEventArgs changeEventArgs)
+    {
+        if (changeEventArgs.Value is not string text)
+        {
+            throw new ArgumentException("Input type was not string", nameof(changeEventArgs));
         }
+
+        return text.Replace(" ", "");
     }
 }
