@@ -1,11 +1,10 @@
-namespace Spreadsheet.Web.WebApp
+namespace Spreadsheet.Web.WebApp;
+
+public static class Layout
 {
-    public static class Layout
-    {
-        public const string FormRow = "row form-row mt-1";
-        public const string LabelContainer = "col-2";
-        public const string Label = "col-form-label";
+    public static readonly string FormRow = "row form-row mt-1";
+    public static readonly string LabelContainer = "col-2";
+    public static readonly string Label = "col-form-label";
         
-        public const string InputContainer = "col-6";
-    }
+    public static readonly string InputContainer = "col-6";
 }
